+++ Engine/Renderer/GLES3/src/GLES3Mesh.cpp
@@ -325,7 +325,7 @@ uint32_t CGLES3Mesh::GetDrawCommandOffset(int indexDraw) const
 
 void CGLES3Mesh::Bind(void *pParam)
 {
-	int indexDraw = (int)pParam;
+	uintptr_t indexDraw = (uintptr_t)pParam;
 
 	if (indexDraw >= 0 && indexDraw < (int)m_draws.size()) {
 		m_draws[indexDraw].pVertexArrayObject->Bind(nullptr);
